Added the virtual environment to the repo
Moved all models to models/ directory.

Changed name of game with gaps

Added dqn_agent_manager.py, and changed achtung.py to use it.

Created a models file (not finished it) and found that Sequential and
Model are both instances of keras.engine.functional.Functional

Removed venv from the repo